benchmark code improved

+++ test/test_utils.py
@@ -10,6 +10,7 @@ import subprocess
 import time
 import wget
 import sys
+import logging
 
 
 
@@ -31,9 +32,63 @@ def get_fastafs_rev():
     return subprocess.check_output(['./bin/fastafs', '--version']).decode("utf-8").split("\n")[0].replace('fastafs','').strip()
 
 
-#git = get_git_revision()
-#print(get_curtime() + "\t" + get_machine_id() + "\t" + git[0] + "\t" + git[1] )
+def check_cache_file_plain(fastafs_binary, filename_in, filename_out):
+    logging.info("Checking if plain fastafs file exists: " + filename_out)
+    
+    if not os.path.exists(filename_out):
+        logging.info("Creating new fastafs archive: " + filename_in)
+        cmd = [fastafs_binary, 'cache', '--fastafs-only', '-o', filename_out, filename_in]
 
+        logging.info("Running: " + ' '.join(cmd))
+        p = subprocess.Popen(cmd,  stderr=subprocess.PIPE, stdout = subprocess.PIPE)#:# as p:
+        stdout, stderr = p.communicate()
+
+
+def check_cache_file_zstd(fastafs_binary, filename_in, filename_out):
+    logging.info("Checking if ZSTD fastafs file exists: " + filename_out)
+    
+    if not os.path.exists(filename_out):
+        logging.info("Creating new zstd archive" + filename_in)
+        
+        cmd = [fastafs_binary, 'cache', '-o', re.sub('.zst$', '', filename_out) , filename_in]
+        logging.info("Running: " + ' '.join(cmd))
+        p = subprocess.Popen(cmd,  stderr=subprocess.PIPE, stdout = subprocess.PIPE)#:# as p:
+        stdout, stderr = p.communicate()
+
+
+def check_benchmark_file(RESULTS_FILE):
+    logging.info("Checking presence of file: " + RESULTS_FILE)
+    
+    if not os.path.exists(RESULTS_FILE):
+        logging.info("Creating benchmark file: " + RESULTS_FILE)
+        with open(RESULTS_FILE,'w') as fh:
+            fh.write(
+                "\t".join(
+                ["timestamp", "fastafs-version", "git-commit", "perf:cycles", "perf:total_time", "perf:user_time", "perf:sys_time", "cmd", "git-mod-status"]
+                ) + "\n"
+            )
+    
+
+def write_benchmark_file(RESULTS_FILE, difference, TIMESTAMP, FASTAFS_REV, GIT_REV):
+    with open(RESULTS_FILE, 'a') as fh:
+        logging.info("Checking if a diff was noted")
+        #print(difference)
+        if(difference['diff']):
+            raise Exception("ERROR - DIFFERENCE DETECTED")
+            import sys
+            sys.exit(1)
+        else:
+            fh.write(
+                "\t".join([TIMESTAMP,
+                           FASTAFS_REV,
+                           GIT_REV[0].replace('git-commit:',''),
+                           str(difference['perf']['cycles']),
+                           str(difference['perf']['total_time']),
+                           str(difference['perf']['user_time']),
+                           str(difference['perf']['sys_time']),
+                           " ".join(difference['cmd']),
+                           GIT_REV[1]]) + "\n"
+            )
 
 
 
@@ -118,22 +173,23 @@ def diff_fasta_with_mounted(fasta_file, fastafs_tmp_filename, fastafs_mount_alia
 
     # 1. fasta to FASTAFS:
     prog = 'cache'
-    print(fastafs_tmp_filename)
     if use_zstd:
         cmd = [fastafs_binary, prog, '-o', fastafs_tmp_filename, fasta_file]
         fastafs_tmp_filename = fastafs_tmp_filename + ".zst"
     else:
         fastafs_tmp_filename = fastafs_tmp_filename + ".fastafs"
         cmd = [fastafs_binary, prog, '--fastafs-only', '-o', fastafs_tmp_filename, fasta_file]
-    print(fastafs_tmp_filename)
 
 
-    #p = subprocess.Popen(cmd,  stderr=subprocess.PIPE, stdout = subprocess.PIPE)#:# as p:
-    #stdout, stderr = p.communicate()
-    #output['cmd'][prog] = cmd
-    #output['stdout'][prog] = stdout.decode("utf-8")
-    #output['stderr'][prog] = stderr.decode("utf-8")
-    #output['retcode'][prog] = p.returncode
+    if not os.path.exists(fastafs_tmp_filename):
+        print(' '.join(cmd))
+
+        p = subprocess.Popen(cmd,  stderr=subprocess.PIPE, stdout = subprocess.PIPE)#:# as p:
+        stdout, stderr = p.communicate()
+        output['cmd'][prog] = cmd
+        output['stdout'][prog] = stdout.decode("utf-8")
+        output['stderr'][prog] = stderr.decode("utf-8")
+        output['retcode'][prog] = p.returncode
 
 
     # 2. check integrity:
@@ -172,10 +228,6 @@ def diff_fasta_with_mounted(fasta_file, fastafs_tmp_filename, fastafs_mount_alia
 
     matches = 0
     try:
-        #os.system('ls -als tmp/benchmark/mnt/*.fa')
-        #time.sleep(0.1)
-        #os.system('ls -als tmp/mnt/*.fa')
-
         with open(fn) as fh_mnt, open(fasta_file) as fh_orig:
             for line1, line2 in tqdm(zip(fh_mnt, fh_orig)):
                 if line1.strip() != line2.strip():
@@ -186,12 +238,6 @@ def diff_fasta_with_mounted(fasta_file, fastafs_tmp_filename, fastafs_mount_alia
                     matches += 1
 
     except Exception as e:
-        # print('-> ',fn)
-        # print(':: ',os.path.exists(fn))
-        # print('-> ',fasta_file)
-        # print(':: ',os.path.exists(fasta_file))
-        # print("***********************************")
-        # print("xxxxx ",str(e))
         output['diff'] = True
 
 
@@ -226,7 +272,7 @@ def diff_fasta_with_mounted(fasta_file, fastafs_tmp_filename, fastafs_mount_alia
 
 
 
-def diff_fasta_with_view(fasta_file, fastafs_tmp_name, padding, fastafs_binary, mountpoint):
+def diff_fasta_with_view(fasta_file, fastafs_tmp_filename, padding, fastafs_binary, use_zstd, mountpoint = "tmp/mnt"):
     """
     Do a diff with an original fasta with 'fastafs view' command
     """
@@ -243,31 +289,40 @@ def diff_fasta_with_view(fasta_file, fastafs_tmp_name, padding, fastafs_binary, 
 
     # 1. fasta to FASTAFS:
     prog = 'cache'
-    #cmd = [fastafs_binary, prog, '-o', fastafs_tmp_file , fasta_file]
-    cmd = [fastafs_binary, prog, fastafs_tmp_name , fasta_file]
-    print(cmd)
-    p = subprocess.Popen(cmd,  stderr=subprocess.PIPE, stdout = subprocess.PIPE)
-    stdout, stderr = p.communicate()
-    output['cmd'][prog] = cmd
-    output['stdout'][prog] = stdout.decode("utf-8")
-    output['stderr'][prog] = stderr.decode("utf-8")
-    output['retcode'][prog] = p.returncode
-    p.terminate()
+    if use_zstd:
+        cmd = [fastafs_binary, prog, '-o', fastafs_tmp_filename, fasta_file]
+        fastafs_tmp_filename = fastafs_tmp_filename + ".zst"
+    else:
+        fastafs_tmp_filename = fastafs_tmp_filename + ".fastafs"
+        cmd = [fastafs_binary, prog, '--fastafs-only', '-o', fastafs_tmp_filename, fasta_file]
     
-    # 2. check integrity:
-    prog = 'check'
-    cmd = [fastafs_binary, 'check', fastafs_tmp_name, fastafs_tmp_name]
-    p = subprocess.Popen(cmd,  stderr=subprocess.PIPE, stdout = subprocess.PIPE)
-    stdout, stderr = p.communicate()
-    output['cmd'][prog] = cmd
-    output['stdout'][prog] = stdout.decode("utf-8")
-    output['stderr'][prog] = stderr.decode("utf-8")
-    output['retcode'][prog] = p.returncode
-    p.terminate()
+    
+    if not os.path.exists(fastafs_tmp_filename):
+        print(' '.join(cmd))
 
-    # 3. run view
+        p = subprocess.Popen(cmd,  stderr=subprocess.PIPE, stdout = subprocess.PIPE)
+        stdout, stderr = p.communicate()
+        output['cmd'][prog] = cmd
+        output['stdout'][prog] = stdout.decode("utf-8")
+        output['stderr'][prog] = stderr.decode("utf-8")
+        output['retcode'][prog] = p.returncode
+        p.terminate()
+    
+    # # 2. check integrity:
+    # prog = 'check'
+    # cmd = [fastafs_binary, 'check', fastafs_tmp_name, fastafs_tmp_name]
+    # p = subprocess.Popen(cmd,  stderr=subprocess.PIPE, stdout = subprocess.PIPE)
+    # stdout, stderr = p.communicate()
+    # output['cmd'][prog] = cmd
+    # output['stdout'][prog] = stdout.decode("utf-8")
+    # output['stderr'][prog] = stderr.decode("utf-8")
+    # output['retcode'][prog] = p.returncode
+    # p.terminate()
+
+    # 3. run view and test outcome
     prog = 'view'
-    cmd = [fastafs_binary, prog, '-p', str(padding), fastafs_tmp_name]
+    cmd = [fastafs_binary, prog, '-p', str(padding), '-f' , fastafs_tmp_filename]
+    print(' '.join(cmd))
     
     os.environ['PYTHONUNBUFFERED'] = "1"
     p = subprocess.Popen(cmd,  stdout=subprocess.PIPE, bufsize=10) # bufsize=1, universal_newlines=True
@@ -284,6 +339,10 @@ def diff_fasta_with_view(fasta_file, fastafs_tmp_name, padding, fastafs_binary, 
                     break
 
     p.stdout.close()
+
+
+    # 4. perf view
+    
     
 
     return output
